third_party/blink/web_tests/virtual/controls-refresh/time-picker/time-picker-cancel-select-value-with-keyboard.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../fast/forms/resources/common.js"></script>
<script src="../../../fast/forms/resources/picker-common.js"></script>
<script src="../../../fast/forms/calendar-picker/resources/calendar-picker-common.js"></script>
</head>
<body>
<input type="time" id="time" value="14:15">
<script>

let t = async_test('Test cancel select value with keyboard in time popup');

function cancelSelectValue() {
  // change hours value
  eventSender.keyDown('ArrowDown');
  eventSender.keyDown('ArrowDown');

  // move to minutes column
  eventSender.keyDown('ArrowRight');
  // change minutes value
  eventSender.keyDown('ArrowDown');
  eventSender.keyDown('ArrowDown');
  eventSender.keyDown('ArrowDown');

  // move to seconds column
  eventSender.keyDown('ArrowRight');
  // change seconds value
  eventSender.keyDown('ArrowUp');
  eventSender.keyDown('ArrowUp');
  eventSender.keyDown('ArrowUp');
  eventSender.keyDown('ArrowUp');
  eventSender.keyDown('ArrowUp');
  eventSender.keyDown('ArrowUp');

  // move to ampm column
  eventSender.keyDown('ArrowRight');
  // change ampm value
  eventSender.keyDown('ArrowDown');

  // close the popup
  eventSender.keyDown('Escape');

  let timeElement = document.getElementById("time");
  assert_equals(timeElement.value, "14:15");
  assert_equals(internals.pagePopupWindow, null);

  t.done();
}

t.step(() => {
  openPicker(document.getElementById('time'), t.step_func(cancelSelectValue),  t.step_func_done(() => {
    assert_false(internals.runtimeFlags.formControlsRefreshEnabled, "Popup did not open.");
  }));
});
</script>
</body>
</html>
